+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-11-05
+
+### Fixed
+- Corrected documentation formatting and improved examples in `lib/yookassa.ex`
+
+### Changed
+- Updated version to 0.2.0 and made `plug_cowboy` dependency optional
+
 ## [0.1.3] - 2025-11-05
 
 ### Changed